debugged delete function, added delete error code

+++ Controllers/PostController.php
@@ -123,31 +123,35 @@ class PostController extends Controller
     
   $confirm = '';     
   if(!empty($query)){
-     /* need to debug this part */
      $slug = '/post/index/' . trim($query);
-     $post = $this->_model->read_post($slug);
+     if ($post = $this->_model->read_post($slug)){
+        $post = $this->_model->read_post($slug);
+        print_r($post);
 
-     $title = $post[0]['title'];
-     $body = $post[0]['body'];
+        $title = $post[0]['title'];
+        $body = $post[0]['body'];
 
-     $this->_view->set('title', $title);
-     $this->_view->set('body', $body );
+        $this->_view->set('title', $title);
+        $this->_view->set('body', $body );
+        
      
-     
-     if ($post[0]['user_username'] === $this->username) { 
-       $confirm = 'Are you sure you want to delete this post?'; 
-       $this->_view->set('confirm',$confirm);
+        if ($post[0]['user_username'] === $this->username) { 
+          $confirm = 'Are you sure you want to delete this post?'; 
 
-       if($_SERVER["REQUEST_METHOD"] == "POST"){ 
+          if($_SERVER["REQUEST_METHOD"] == "POST"){ 
            $this->_model->delete_post($slug);
-       }
-      } else {
+          }
+        } else {
                   
-         $error = "Only authorized users can delete a post";
-                }
-      } // end of username auth block        
+          $error = "Only authorized users can delete a post";
+         } // end of auth user check 
+       } else{
+         $error = 'This post does not exist';
+       }// end of if post exists check block
+      } // end of if empty query check        
      
    $this->_view->set('error', $error);
+   $this->_view->set('confirm', $confirm);
    return $this->_view->render();
 } // end of delete post function
